컴포넌트 분리 및 아파트 평점 디자인 작업
