+++ src/components/main/details/HouseDetailHeader.tsx
@@ -1,0 +1,28 @@
+import "./style/detail.css";
+import React from "react";
+import { HouseDetailHeaderStyle, HouseHeader } from "../style/main-item.style";
+import { ReactComponent as HouseCloseIcon } from "../../../assets/icon/close.svg";
+import { ReactComponent as HeartIcon } from "../../../assets/icon/heart.svg";
+import { useNavigate } from "react-router-dom";
+
+const HouseDetailHeader: React.FC = () => {
+    const navigate = useNavigate();
+
+    return (
+        <HouseDetailHeaderStyle>
+            <HouseHeader>
+                아파트 정보
+                <HouseCloseIcon className="close" onClick={() => navigate(-1)} />
+            </HouseHeader>
+            <div className="image-container">
+                <img src="" alt="preview" className="preview-image"/>
+            </div>
+            <div className="interested">
+                <HeartIcon />
+                &nbsp; 관심 매물 추가하기
+            </div>
+        </HouseDetailHeaderStyle> 
+    )
+}
+
+export default HouseDetailHeader;
