+++ src/components/main/style/main.css
@@ -190,120 +190,3 @@
     height: 100vh;
     z-index: -1;
 }
-
-.item {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-around;
-}
-
-.item span {
-    margin: 0 10px 0 5px;
-    font-size: 12px;
-    color: #03C6CE;
-}
-
-.item .info {
-    display: flex;
-    flex-direction: column;
-}
-
-.item .rank {
-    font-size: 14px;
-    color: #03C6CE;
-}
-
-.item .info .address {
-    font-size: 12px;
-    color: #9DA3AF;
-}
-
-.item .info span {
-    font-size: 14px;
-    color: #111;
-}
-
-.item .price-info {
-    display: flex;
-    flex-direction: column;
-    align-items: end;
-}
-
-.chip-list {
-    display: flex;
-    gap: 10px;
-    padding: 10px;
-    max-width: 100%;
-    overflow-x: auto;
-}
-
-.chip {
-    display: inline-block;
-    padding: 10px 20px;
-    border: 1px solid #ccc;
-    border-radius: 25px;
-    cursor: pointer;
-    transition: background-color 0.3s, border-color 0.3s;
-    white-space: nowrap; /* Prevents text from wrapping */
-}
-
-.chip:hover {
-    border-color: #aaa;
-}
-
-.chip.active {
-    background-color: #1abc9c;
-    color: white;
-    border: none;
-}
-
-.chip:not(.active) {
-    background-color: white;
-    color: black;
-}
-
-.radios {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-around;
-}
-
-.comment-list {
-    display: flex;
-    flex-direction: column;
-}
-
-.comment-list .comment-item {
-    width: 100%;
-    height: 80px;
-    margin: 10px 0 10px 0;
-}
-
-.comment-item .user {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-}
-
-.comment-item .user .data {
-    display: flex;
-    flex-direction: column;
-    align-items: start;
-}
-
-.comment-item .user .icon {
-    width: 40px;
-    height: 40px;
-    border-radius: 100%;
-    background-color: #D9D9D9;
-    margin-right: 10px;
-}
-
-.comment-item .user .data .nickname {
-    font-size: 14px;
-}
-
-.comment-item .user .data .date {
-    font-size: 10px;
-    color: #9DA3AF;
-}
